pyproject.toml
[project]
name = "kicad-generator"
version = "0.1.0"
description = "Automation tools that generate KiCAD footprints and symbols for SiFli chips/modules."
readme = "README.md"
authors = [
    { name = "HalfSweet", email = "halfsweet@halfsweet.cn" }
]
requires-python = ">=3.13"
license = { text = "Apache-2.0" }
dependencies = [
    "pyyaml>=6.0.2",
    "typing-extensions>=4.12.2",
    "asteval>=0.9.33",
    "pyclipper>=1.3.0.post6",
]

[project.scripts]
kicad-generator = "kicad_generator.cli:main"

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"
